feat(): Implemented "POST" and "PUT" requests for events
"POST" to "/events" will take in a JSON body with the keys and values corresponding to the event

parameters, create the event in the server, and then respond with a JSON of that event. "PUT" to

"/events/[:id]" will take in a JSON body just like the "POST", except it only needs the parameters

that have been changed. It will also respond with a JSON of the event that was modified.